Add Polyfill for Scheduler API
And update to match as close to spec as possible.

Does not include `TaskSignal` or `TaskController` since changing task priority doesn't seem feasible/worth the effort for what seems to be an edge case.
